requirements.txt
modbus-tk==1.1.2
pyserial==3.5
